Change how errno values are printed, for portability
Reported and suggested by Pino Toscano: The Hurd does not #define the
EFOO constants as simple integers, but as expressions. This made the
code that generates errnos.h to misinterpret the value, and always print
0 for every error code. We fix this by instead using the EFOO macro in
errnos.h, so that the preprocessor evaluates it correctly. This is the
right thing to do anyway, for every platform.

Thus the table changes from {"EFOO",42} to {"EFOO",EFOO} for each
entry.