vidir handle items with common prefix (moreutils debian bug 654326)
Hi Joey,

Many thanks for your work for debian over the years, and I wish you
plenty of energy to continue work on any other projects that you find
useful.

Here's a small patch that fixes debian bug 654326.  vidir modifies its
in-memory copy of filenames in a way that can clobber items that have
common prefix.  There's a one-liner testcase in the bug report.  This
patch fixes it so regexp replacement is anchored at end of line or path
component.